Fixes
Fixes for gcc and intel compilers.
gcc-4.1 clean with -Wall -Wextra
intel 11.1 clean with -Wcheck -w1
msvc2010 gives one unrelated warning

+++ timer.cpp
@@ -1,22 +1,23 @@
+
 #if defined(_WIN32)
 #include <windows.h>
-	long double Timer()
+	double Timer()
 	{
 		LARGE_INTEGER ticksPerSecond;
 		LARGE_INTEGER tick;
 		QueryPerformanceFrequency(&ticksPerSecond);
 		QueryPerformanceCounter(&tick);
-		return (long double)tick.QuadPart/(long double)ticksPerSecond.QuadPart;
+		return (double)tick.QuadPart/(double)ticksPerSecond.QuadPart;
 	}
 #else
 #include <sys/time.h>
 #include <stddef.h>
 	long double Timer()
 	{
-		long double t1;
+		double t1;
 		struct timeval time;
 		gettimeofday(&(time), NULL);
-		t1 =  (long double)time.tv_sec + (long double)time.tv_usec/(1000.0*1000.0);
+		t1 =  (double)time.tv_sec + (double)time.tv_usec/(1000.0*1000.0);
 		return t1;
 	}
 #endif
